742% reduction in image size!
I guess cargo includes a bunch of stuff in target/release that is just
there so you don't have to recompile everything all the time, and you
don't actually need it to run. Now instead of compiling all of
target/release to the image I am only copying the executable. I also had
it strip the symbols from the elf when compiling with the release
profile.